Character Controller with rotate, Lists, and Applying force attempt 1
Added Rotate for 3D space with Character controller. Thinking about utilizing the "do work" in the MatchID script with Searching Inventory. Too narrow?
Started Inventory with Lists, unsure if this is the right path to go down for this. Attempting to create efficient list types for game.
Apply Force script did not entirely work for the first attempt. It seems that only ridgedbodies make this work. More research needed to figure out how to take ridgedbody out of the solution so it doesn't battle for dominance with player.